feat: add InvalidIdException for delete in EmployeeController

+++ src/main/java/com/rest/springbootemployee/controller/EmployeeController.java
@@ -64,7 +64,10 @@ public class EmployeeController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable String id) {
+    public void delete(@NonNull @PathVariable String id) {
+        if(!ObjectId.isValid(id)){
+            throw new InvalidIdException();
+        }
         employeeService.delete(id);
     }
